+++ RichMessageKit/Views/core/SuggestedReplyView.swift
@@ -144,7 +144,12 @@ public class SuggestedReplyView: UIView {
                 let hiddenView = hiddenViewUsing(currWidth: width - button.buttonWidth(), maxWidth: maxWidth, subViews: subviews)
                 alignLeft ? subviews.append(hiddenView) : subviews.insert(hiddenView, at: 0)
                 width = button.buttonWidth() + 10
-                let stackView = horizontalStackView(subviews: subviews)
+                let stackView: UIStackView
+                if subviews.count > 2 {
+                    stackView = horizontalStackViewWithSpace(subviews: subviews)
+                } else {
+                    stackView = horizontalStackView(subviews: subviews)
+                }
                 mainStackView.addArrangedSubview(stackView)
                 subviews.removeAll()
                 subviews.append(button)
@@ -157,12 +162,25 @@ public class SuggestedReplyView: UIView {
         if isMultiLine {
             alignLeft ? subviews.append(hiddenView) : subviews.insert(hiddenView, at: 0)
         }
-
-        let stackView = horizontalStackView(subviews: subviews)
+        let stackView: UIStackView
+        if subviews.count > 2 {
+            stackView = horizontalStackViewWithSpace(subviews: subviews)
+        } else {
+            stackView = horizontalStackView(subviews: subviews)
+        }
+        
         mainStackView.addArrangedSubview(stackView)
     }
 
     private func horizontalStackView(subviews: [UIView]) -> UIStackView {
+        let stackView = UIStackView(arrangedSubviews: subviews)
+        stackView.alignment = .center
+        stackView.axis = .horizontal
+        stackView.distribution = .fill
+        return stackView
+    }
+    
+    private func horizontalStackViewWithSpace(subviews: [UIView]) -> UIStackView {
         let stackView = UIStackView(arrangedSubviews: subviews)
         stackView.spacing = 10
         stackView.alignment = .center