+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
+## Unreleased
+- Dark Mode Support Enabled.
+
 ## [1.1.8] 2023-09-26
 - Added Support For Auto Suggestions Rich Message
 - Added custom input field rich message support in IOS SDK