+++ CHANGELOG.md
@@ -10,11 +10,13 @@ The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/
 - Added support for XCode 15
 - Fixed kommunicate.io showing in iOS 17 even when it is disabled
 
+
 ## [1.1.7] 2023-09-07
 - Added support for Sending GIF from device
 - Added icons for (mobile/web/facebook/WhatsApp) on conversation list for agent app. 
 - Exposed a customisation function for a rating menu icon on conversation screen.
 - Fixed crash coming in agent app due to moreBar.
+
 
 ## [1.1.6] 2023-08-28
 - Fixed attachments upload issue