+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
+##Unreleased
+- Fixed conversation history mismatch
+
 ## [1.1.8] 2023-09-26
 - Added Support For Auto Suggestions Rich Message
 - Added custom input field rich message support in IOS SDK
@@ -9,7 +12,6 @@ The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/
 - Fixed the delete conversation showing the empty screen.  
 - Added support for XCode 15
 - Fixed kommunicate.io showing in iOS 17 even when it is disabled
-
 
 ## [1.1.7] 2023-09-07
 - Added support for Sending GIF from device