+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
 ## Unreleased
+- Fixed the Button Title hidding in smaller devices.
 - Changed foreground color for link present inside message
 
 ## [1.1.8] 2023-09-26