+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
-## Unreleased
+## [1.1.9] 2023-10-18
 - Fixed the Button Title hidding in smaller devices.
 - Changed foreground color for link present inside message.
 - Fixed all messages are not loading in conversation screen.